update test vault generation

+++ make-test-vault.sh
@@ -1,9 +1,13 @@
 #!/usr/bin/env bash
 
-set -x # echo on
+[[ -f ./dist/main.js && -f ./dist/styles.css && -f ./dist/manifest.json ]] || { echo "No build found."; exit 1; };
 
-bun ./build.cjs
-mkdir -p ./test-vault/.obsidian/plugins/obsidian-typewriter-mode
-echo '["typewriter-mode"]' > ./test-vault/.obsidian/community-plugins.json
-cp -r dist/* ./test-vault/.obsidian/plugins/obsidian-typewriter-mode
+echo "Creating test vault.";
+mkdir -p ./test-vault/.obsidian/plugins/obsidian-typewriter-mode;
+[[ -f ./test-vault/.obsidian/community-plugins.json ]] || { echo '["typewriter-mode"]' > ./test-vault/.obsidian/community-plugins.json; };
+
+echo "Copying build to test vault.";
+cp -r dist/* ./test-vault/.obsidian/plugins/obsidian-typewriter-mode;
+
+echo "Test vault successfully prepared.";
